fixes after review

+++ resources/views/shared/simple-resource-index.blade.php
@@ -2,11 +2,11 @@
     <div class="mb-5 flex flex-wrap items-center justify-between gap-4">
         <h1 class="text-4xl md:text-4xl xl:text-5xl">{{ $title }}</h1>
 
-        @auth
+        @can('create', $modelClass)
             <a href="{{ $createRoute }}" class="app-button shrink-0">
                 {{ $createLabel }}
             </a>
-        @endauth
+        @endcan
     </div>
 
     <div class="page-card overflow-x-auto">
@@ -19,9 +19,9 @@
                         <th class="px-3 py-2">{{ __('strings.description') }}</th>
                     @endif
                     <th class="whitespace-nowrap px-3 py-2">{{ __('strings.data created') }}</th>
-                    @auth
+                    @can('create', $modelClass)
                         <th class="whitespace-nowrap px-3 py-2">{{ __('strings.actions') }}</th>
-                    @endauth
+                    @endcan
                 </tr>
             </thead>
             <tbody>
@@ -33,24 +33,28 @@
                             <td class="px-3 py-3">{{ $item->description }}</td>
                         @endif
                         <td class="whitespace-nowrap px-3 py-3">{{ $item->created_at->format('d.m.Y') }}</td>
-                        @auth
+                        @can('create', $modelClass)
                             <td class="whitespace-nowrap px-3 py-3">
-                                @if ($item->tasks_count === 0)
+                                @can('delete', $item)
+                                    @if ($item->tasks_count === 0)
+                                        <a
+                                            data-method="delete"
+                                            data-confirm="{{ __('strings.are you sure') }}"
+                                            class="text-red-600 hover:text-red-900"
+                                            href="{{ route($destroyRouteName, $item) }}"
+                                        >{{ __('strings.delete') }}</a>
+                                    @elseif (! empty($inUseHint))
+                                        <span class="text-sm text-slate-500">{{ $inUseHint }}</span>
+                                    @endif
+                                @endcan
+                                @can('update', $item)
                                     <a
-                                        data-method="delete"
-                                        data-confirm="{{ __('strings.are you sure') }}"
-                                        class="text-red-600 hover:text-red-900"
-                                        href="{{ route($destroyRouteName, $item) }}"
-                                    >{{ __('strings.delete') }}</a>
-                                @elseif (! empty($inUseHint))
-                                    <span class="text-sm text-slate-500">{{ $inUseHint }}</span>
-                                @endif
-                                <a
-                                    class="ml-3 text-blue-600 hover:text-blue-900"
-                                    href="{{ route($editRouteName, $item) }}"
-                                >{{ __('strings.edit') }}</a>
+                                        class="ml-3 text-blue-600 hover:text-blue-900"
+                                        href="{{ route($editRouteName, $item) }}"
+                                    >{{ __('strings.edit') }}</a>
+                                @endcan
                             </td>
-                        @endauth
+                        @endcan
                     </tr>
                 @endforeach
             </tbody>
